WAR: ветка веб-приложения

+++ src/main/java/com/century/report/generator/ExportSalesReportGenerator.java
@@ -8,23 +8,20 @@ import java.io.File;
 import java.math.BigDecimal;
 import java.util.List;
 
-import static com.century.report.Util.getExcelFileFullPath;
-import static com.century.report.Util.logToFile;
+import static com.century.report.service.util.UtilService.getResultFileName;
 import static java.math.RoundingMode.HALF_UP;
 
 public abstract class ExportSalesReportGenerator <T> implements ReportGenerator {
-     protected  ReportType reportType;
-     protected ReportSettings settings;
-     protected List<T> data;
+    protected  ReportType reportType;
+    protected ReportSettings settings;
+    protected List<T> data;
 
     public ExportSalesReportGenerator(ReportType reportType, ReportSettings settings, List<T> data){
         this.reportType = reportType;
         this.settings = settings;
         this.data = data;
 
-        if(!deleteOutputFileIfExists()){
-            logToFile(settings.getUsername(), "Failed to delete output file.");
-        }
+        setFileName();
         verifyInput();
         verifyReportType();
     }
@@ -33,10 +30,24 @@ public abstract class ExportSalesReportGenerator <T> implements ReportGenerator 
         return input.setScale(settings.getDecimalPlaces(), HALF_UP);
     }
 
-    private boolean deleteOutputFileIfExists(){
+    protected abstract void logToFile(String username, String msg);
+
+    private void setFileName(){
+        String filename = getResultFileName();
+
+        if(!deleteOutputFileIfExists(filename)){
+            logToFile(settings.getUsername(), "Failed to delete output file.");
+        }
+        this.settings.setFilename(filename);
+    }
+
+    private boolean deleteOutputFileIfExists(String filename){
         try{
-            File existing = new File(getExcelFileFullPath(settings.getFilename()));
-            return existing.delete();
+            File existing = new File(getExcelFileFullPath(filename));
+            if(existing.exists()){
+                return existing.delete();
+            }
+            return true;
         } catch (Exception e) {
             //NOP
         }
@@ -61,4 +72,6 @@ public abstract class ExportSalesReportGenerator <T> implements ReportGenerator 
             throw new ExportSalesReportException("Не заданы входные данные для отчёта.");
         }
     }
+
+    protected abstract String getExcelFileFullPath(String filename);
 }
